Struct aws_sdk_cloudfront::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Cloudfront2020_05_31.

This client allows ergonomic access to a Cloudfront2020_05_31-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_alias(&self) -> AssociateAlias<C, M, R>

pub fn create_cache_policy(&self) -> CreateCachePolicy<C, M, R>

pub fn create_cloud_front_origin_access_identity(
&self
) -> CreateCloudFrontOriginAccessIdentity<C, M, R>

pub fn create_distribution(&self) -> CreateDistribution<C, M, R>

pub fn create_distribution_with_tags(
&self
) -> CreateDistributionWithTags<C, M, R>

pub fn create_field_level_encryption_config(
&self
) -> CreateFieldLevelEncryptionConfig<C, M, R>

pub fn create_field_level_encryption_profile(
&self
) -> CreateFieldLevelEncryptionProfile<C, M, R>

pub fn create_function(&self) -> CreateFunction<C, M, R>

pub fn create_invalidation(&self) -> CreateInvalidation<C, M, R>

pub fn create_key_group(&self) -> CreateKeyGroup<C, M, R>

pub fn create_monitoring_subscription(
&self
) -> CreateMonitoringSubscription<C, M, R>

pub fn create_origin_request_policy(&self) -> CreateOriginRequestPolicy<C, M, R>

pub fn create_public_key(&self) -> CreatePublicKey<C, M, R>

pub fn create_realtime_log_config(&self) -> CreateRealtimeLogConfig<C, M, R>

pub fn create_streaming_distribution(
&self
) -> CreateStreamingDistribution<C, M, R>

pub fn create_streaming_distribution_with_tags(
&self
) -> CreateStreamingDistributionWithTags<C, M, R>

pub fn delete_cache_policy(&self) -> DeleteCachePolicy<C, M, R>

pub fn delete_cloud_front_origin_access_identity(
&self
) -> DeleteCloudFrontOriginAccessIdentity<C, M, R>

pub fn delete_distribution(&self) -> DeleteDistribution<C, M, R>

pub fn delete_field_level_encryption_config(
&self
) -> DeleteFieldLevelEncryptionConfig<C, M, R>

pub fn delete_field_level_encryption_profile(
&self
) -> DeleteFieldLevelEncryptionProfile<C, M, R>

pub fn delete_function(&self) -> DeleteFunction<C, M, R>

pub fn delete_key_group(&self) -> DeleteKeyGroup<C, M, R>

pub fn delete_monitoring_subscription(
&self
) -> DeleteMonitoringSubscription<C, M, R>

pub fn delete_origin_request_policy(&self) -> DeleteOriginRequestPolicy<C, M, R>

pub fn delete_public_key(&self) -> DeletePublicKey<C, M, R>

pub fn delete_realtime_log_config(&self) -> DeleteRealtimeLogConfig<C, M, R>

pub fn delete_streaming_distribution(
&self
) -> DeleteStreamingDistribution<C, M, R>

pub fn describe_function(&self) -> DescribeFunction<C, M, R>

pub fn get_cache_policy(&self) -> GetCachePolicy<C, M, R>

pub fn get_cache_policy_config(&self) -> GetCachePolicyConfig<C, M, R>

pub fn get_cloud_front_origin_access_identity(
&self
) -> GetCloudFrontOriginAccessIdentity<C, M, R>

pub fn get_cloud_front_origin_access_identity_config(
&self
) -> GetCloudFrontOriginAccessIdentityConfig<C, M, R>

pub fn get_distribution(&self) -> GetDistribution<C, M, R>

pub fn get_distribution_config(&self) -> GetDistributionConfig<C, M, R>

pub fn get_field_level_encryption(&self) -> GetFieldLevelEncryption<C, M, R>

pub fn get_field_level_encryption_config(
&self
) -> GetFieldLevelEncryptionConfig<C, M, R>

pub fn get_field_level_encryption_profile(
&self
) -> GetFieldLevelEncryptionProfile<C, M, R>

pub fn get_field_level_encryption_profile_config(
&self
) -> GetFieldLevelEncryptionProfileConfig<C, M, R>

pub fn get_function(&self) -> GetFunction<C, M, R>

pub fn get_invalidation(&self) -> GetInvalidation<C, M, R>

pub fn get_key_group(&self) -> GetKeyGroup<C, M, R>

pub fn get_key_group_config(&self) -> GetKeyGroupConfig<C, M, R>

pub fn get_monitoring_subscription(&self) -> GetMonitoringSubscription<C, M, R>

pub fn get_origin_request_policy(&self) -> GetOriginRequestPolicy<C, M, R>

pub fn get_origin_request_policy_config(
&self
) -> GetOriginRequestPolicyConfig<C, M, R>

pub fn get_public_key(&self) -> GetPublicKey<C, M, R>

pub fn get_public_key_config(&self) -> GetPublicKeyConfig<C, M, R>

pub fn get_realtime_log_config(&self) -> GetRealtimeLogConfig<C, M, R>

pub fn get_streaming_distribution(&self) -> GetStreamingDistribution<C, M, R>

pub fn get_streaming_distribution_config(
&self
) -> GetStreamingDistributionConfig<C, M, R>

pub fn list_cache_policies(&self) -> ListCachePolicies<C, M, R>

pub fn list_cloud_front_origin_access_identities(
&self
) -> ListCloudFrontOriginAccessIdentities<C, M, R>

pub fn list_conflicting_aliases(&self) -> ListConflictingAliases<C, M, R>

pub fn list_distributions(&self) -> ListDistributions<C, M, R>

pub fn list_distributions_by_cache_policy_id(
&self
) -> ListDistributionsByCachePolicyId<C, M, R>

pub fn list_distributions_by_key_group(
&self
) -> ListDistributionsByKeyGroup<C, M, R>

pub fn list_distributions_by_origin_request_policy_id(
&self
) -> ListDistributionsByOriginRequestPolicyId<C, M, R>

pub fn list_distributions_by_realtime_log_config(
&self
) -> ListDistributionsByRealtimeLogConfig<C, M, R>

pub fn list_distributions_by_web_acl_id(
&self
) -> ListDistributionsByWebACLId<C, M, R>

pub fn list_field_level_encryption_configs(
&self
) -> ListFieldLevelEncryptionConfigs<C, M, R>

pub fn list_field_level_encryption_profiles(
&self
) -> ListFieldLevelEncryptionProfiles<C, M, R>

pub fn list_functions(&self) -> ListFunctions<C, M, R>

pub fn list_invalidations(&self) -> ListInvalidations<C, M, R>

pub fn list_key_groups(&self) -> ListKeyGroups<C, M, R>

pub fn list_origin_request_policies(&self) -> ListOriginRequestPolicies<C, M, R>

pub fn list_public_keys(&self) -> ListPublicKeys<C, M, R>

pub fn list_realtime_log_configs(&self) -> ListRealtimeLogConfigs<C, M, R>

pub fn list_streaming_distributions(
&self
) -> ListStreamingDistributions<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn publish_function(&self) -> PublishFunction<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_function(&self) -> TestFunction<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_cache_policy(&self) -> UpdateCachePolicy<C, M, R>

pub fn update_cloud_front_origin_access_identity(
&self
) -> UpdateCloudFrontOriginAccessIdentity<C, M, R>

pub fn update_distribution(&self) -> UpdateDistribution<C, M, R>

pub fn update_field_level_encryption_config(
&self
) -> UpdateFieldLevelEncryptionConfig<C, M, R>

pub fn update_field_level_encryption_profile(
&self
) -> UpdateFieldLevelEncryptionProfile<C, M, R>

pub fn update_function(&self) -> UpdateFunction<C, M, R>

pub fn update_key_group(&self) -> UpdateKeyGroup<C, M, R>

pub fn update_origin_request_policy(&self) -> UpdateOriginRequestPolicy<C, M, R>

pub fn update_public_key(&self) -> UpdatePublicKey<C, M, R>

pub fn update_realtime_log_config(&self) -> UpdateRealtimeLogConfig<C, M, R>

pub fn update_streaming_distribution(
&self
) -> UpdateStreamingDistribution<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self