Module aws_sdk_cloudfront::error

Modules

access_denied

See AccessDenied

batch_too_large

See BatchTooLarge

cache_policy_already_exists

See CachePolicyAlreadyExists

cache_policy_in_use

See CachePolicyInUse

cannot_change_immutable_public_key_fields

See CannotChangeImmutablePublicKeyFields

cloud_front_origin_access_identity_already_exists

See CloudFrontOriginAccessIdentityAlreadyExists

cloud_front_origin_access_identity_in_use

See CloudFrontOriginAccessIdentityInUse

cname_already_exists

See CnameAlreadyExists

distribution_already_exists

See DistributionAlreadyExists

distribution_not_disabled

See DistributionNotDisabled

field_level_encryption_config_already_exists

See FieldLevelEncryptionConfigAlreadyExists

field_level_encryption_config_in_use

See FieldLevelEncryptionConfigInUse

field_level_encryption_profile_already_exists

See FieldLevelEncryptionProfileAlreadyExists

field_level_encryption_profile_in_use

See FieldLevelEncryptionProfileInUse

field_level_encryption_profile_size_exceeded

See FieldLevelEncryptionProfileSizeExceeded

function_already_exists

See FunctionAlreadyExists

function_in_use

See FunctionInUse

function_size_limit_exceeded

See FunctionSizeLimitExceeded

illegal_delete

See IllegalDelete

illegal_field_level_encryption_config_association_with_cache_behavior

See IllegalFieldLevelEncryptionConfigAssociationWithCacheBehavior

illegal_update

See IllegalUpdate

inconsistent_quantities

See InconsistentQuantities

invalid_argument

See InvalidArgument

invalid_default_root_object

See InvalidDefaultRootObject

invalid_error_code

See InvalidErrorCode

invalid_forward_cookies

See InvalidForwardCookies

invalid_function_association

See InvalidFunctionAssociation

invalid_geo_restriction_parameter

See InvalidGeoRestrictionParameter

invalid_headers_for_s3_origin

See InvalidHeadersForS3Origin

invalid_if_match_version

See InvalidIfMatchVersion

invalid_lambda_function_association

See InvalidLambdaFunctionAssociation

invalid_location_code

See InvalidLocationCode

invalid_minimum_protocol_version

See InvalidMinimumProtocolVersion

invalid_origin

See InvalidOrigin

invalid_origin_access_identity

See InvalidOriginAccessIdentity

invalid_origin_keepalive_timeout

See InvalidOriginKeepaliveTimeout

invalid_origin_read_timeout

See InvalidOriginReadTimeout

invalid_protocol_settings

See InvalidProtocolSettings

invalid_query_string_parameters

See InvalidQueryStringParameters

invalid_relative_path

See InvalidRelativePath

invalid_required_protocol

See InvalidRequiredProtocol

invalid_response_code

See InvalidResponseCode

invalid_tagging

See InvalidTagging

invalid_ttl_order

See InvalidTtlOrder

invalid_viewer_certificate

See InvalidViewerCertificate

invalid_web_acl_id

See InvalidWebAclId

key_group_already_exists

See KeyGroupAlreadyExists

missing_body

See MissingBody

no_such_cache_policy

See NoSuchCachePolicy

no_such_cloud_front_origin_access_identity

See NoSuchCloudFrontOriginAccessIdentity

no_such_distribution

See NoSuchDistribution

no_such_field_level_encryption_config

See NoSuchFieldLevelEncryptionConfig

no_such_field_level_encryption_profile

See NoSuchFieldLevelEncryptionProfile

no_such_function_exists

See NoSuchFunctionExists

no_such_invalidation

See NoSuchInvalidation

no_such_origin

See NoSuchOrigin

no_such_origin_request_policy

See NoSuchOriginRequestPolicy

no_such_public_key

See NoSuchPublicKey

no_such_realtime_log_config

See NoSuchRealtimeLogConfig

no_such_resource

See NoSuchResource

no_such_streaming_distribution

See NoSuchStreamingDistribution

origin_request_policy_already_exists

See OriginRequestPolicyAlreadyExists

origin_request_policy_in_use

See OriginRequestPolicyInUse

precondition_failed

See PreconditionFailed

public_key_already_exists

See PublicKeyAlreadyExists

public_key_in_use

See PublicKeyInUse

query_arg_profile_empty

See QueryArgProfileEmpty

realtime_log_config_already_exists

See RealtimeLogConfigAlreadyExists

realtime_log_config_in_use

See RealtimeLogConfigInUse

realtime_log_config_owner_mismatch

See RealtimeLogConfigOwnerMismatch

resource_in_use

See ResourceInUse

streaming_distribution_already_exists

See StreamingDistributionAlreadyExists

streaming_distribution_not_disabled

See StreamingDistributionNotDisabled

test_function_failed

See TestFunctionFailed

too_many_cache_behaviors

See TooManyCacheBehaviors

too_many_cache_policies

See TooManyCachePolicies

too_many_certificates

See TooManyCertificates

too_many_cloud_front_origin_access_identities

See TooManyCloudFrontOriginAccessIdentities

too_many_cookie_names_in_white_list

See TooManyCookieNamesInWhiteList

too_many_cookies_in_cache_policy

See TooManyCookiesInCachePolicy

too_many_cookies_in_origin_request_policy

See TooManyCookiesInOriginRequestPolicy

too_many_distribution_cnam_es

See TooManyDistributionCnamEs

too_many_distributions

See TooManyDistributions

too_many_distributions_associated_to_cache_policy

See TooManyDistributionsAssociatedToCachePolicy

too_many_distributions_associated_to_field_level_encryption_config

See TooManyDistributionsAssociatedToFieldLevelEncryptionConfig

too_many_distributions_associated_to_key_group

See TooManyDistributionsAssociatedToKeyGroup

too_many_distributions_associated_to_origin_request_policy

See TooManyDistributionsAssociatedToOriginRequestPolicy

too_many_distributions_with_function_associations

See TooManyDistributionsWithFunctionAssociations

too_many_distributions_with_lambda_associations

See TooManyDistributionsWithLambdaAssociations

too_many_distributions_with_single_function_arn

See TooManyDistributionsWithSingleFunctionArn

too_many_field_level_encryption_configs

See TooManyFieldLevelEncryptionConfigs

too_many_field_level_encryption_content_type_profiles

See TooManyFieldLevelEncryptionContentTypeProfiles

too_many_field_level_encryption_encryption_entities

See TooManyFieldLevelEncryptionEncryptionEntities

too_many_field_level_encryption_field_patterns

See TooManyFieldLevelEncryptionFieldPatterns

too_many_field_level_encryption_profiles

See TooManyFieldLevelEncryptionProfiles

too_many_field_level_encryption_query_arg_profiles

See TooManyFieldLevelEncryptionQueryArgProfiles

too_many_function_associations

See TooManyFunctionAssociations

too_many_functions

See TooManyFunctions

too_many_headers_in_cache_policy

See TooManyHeadersInCachePolicy

too_many_headers_in_forwarded_values

See TooManyHeadersInForwardedValues

too_many_headers_in_origin_request_policy

See TooManyHeadersInOriginRequestPolicy

too_many_invalidations_in_progress

See TooManyInvalidationsInProgress

too_many_key_groups

See TooManyKeyGroups

too_many_key_groups_associated_to_distribution

See TooManyKeyGroupsAssociatedToDistribution

too_many_lambda_function_associations

See TooManyLambdaFunctionAssociations

too_many_origin_custom_headers

See TooManyOriginCustomHeaders

too_many_origin_groups_per_distribution

See TooManyOriginGroupsPerDistribution

too_many_origin_request_policies

See TooManyOriginRequestPolicies

too_many_origins

See TooManyOrigins

too_many_public_keys

See TooManyPublicKeys

too_many_public_keys_in_key_group

See TooManyPublicKeysInKeyGroup

too_many_query_string_parameters

See TooManyQueryStringParameters

too_many_query_strings_in_cache_policy

See TooManyQueryStringsInCachePolicy

too_many_query_strings_in_origin_request_policy

See TooManyQueryStringsInOriginRequestPolicy

too_many_realtime_log_configs

See TooManyRealtimeLogConfigs

too_many_streaming_distribution_cnam_es

See TooManyStreamingDistributionCnamEs

too_many_streaming_distributions

See TooManyStreamingDistributions

too_many_trusted_signers

See TooManyTrustedSigners

trusted_key_group_does_not_exist

See TrustedKeyGroupDoesNotExist

trusted_signer_does_not_exist

See TrustedSignerDoesNotExist

unsupported_operation

See UnsupportedOperation

Structs

AccessDenied

Access denied.

AssociateAliasError

BatchTooLarge

Invalidation batch specified is too large.

CachePolicyAlreadyExists

A cache policy with this name already exists. You must provide a unique name. To

CachePolicyInUse

Cannot delete the cache policy because it is attached to one or more cache

CannotChangeImmutablePublicKeyFields

You can't change the value of a public key.

CloudFrontOriginAccessIdentityAlreadyExists

If the CallerReference is a value you already sent in a previous request to create an identity but the content

CloudFrontOriginAccessIdentityInUse

The Origin Access Identity specified is already in use.

CnameAlreadyExists

The CNAME specified is already defined for CloudFront.

CreateCachePolicyError

CreateCloudFrontOriginAccessIdentityError

CreateDistributionError

CreateDistributionWithTagsError

CreateFieldLevelEncryptionConfigError

CreateFieldLevelEncryptionProfileError

CreateFunctionError

CreateInvalidationError

CreateKeyGroupError

CreateMonitoringSubscriptionError

CreateOriginRequestPolicyError

CreatePublicKeyError

CreateRealtimeLogConfigError

CreateStreamingDistributionError

CreateStreamingDistributionWithTagsError

DeleteCachePolicyError

DeleteCloudFrontOriginAccessIdentityError

DeleteDistributionError

DeleteFieldLevelEncryptionConfigError

DeleteFieldLevelEncryptionProfileError

DeleteFunctionError

DeleteKeyGroupError

DeleteMonitoringSubscriptionError

DeleteOriginRequestPolicyError

DeletePublicKeyError

DeleteRealtimeLogConfigError

DeleteStreamingDistributionError

DescribeFunctionError

DistributionAlreadyExists

The caller reference you attempted to create the distribution with is associated with another distribution.

DistributionNotDisabled

The specified CloudFront distribution is not disabled. You must disable

FieldLevelEncryptionConfigAlreadyExists

The specified configuration for field-level encryption already exists.

FieldLevelEncryptionConfigInUse

The specified configuration for field-level encryption is in use.

FieldLevelEncryptionProfileAlreadyExists

The specified profile for field-level encryption already exists.

FieldLevelEncryptionProfileInUse

The specified profile for field-level encryption is in use.

FieldLevelEncryptionProfileSizeExceeded

The maximum size of a profile for field-level encryption was exceeded.

FunctionAlreadyExists

A function with the same name already exists in this account. To create a

FunctionInUse

Cannot delete the function because it’s attached to one or more cache

FunctionSizeLimitExceeded

The function is too large. For more information, see Quotas (formerly known as limits) in the

GetCachePolicyConfigError

GetCachePolicyError

GetCloudFrontOriginAccessIdentityConfigError

GetCloudFrontOriginAccessIdentityError

GetDistributionConfigError

GetDistributionError

GetFieldLevelEncryptionConfigError

GetFieldLevelEncryptionError

GetFieldLevelEncryptionProfileConfigError

GetFieldLevelEncryptionProfileError

GetFunctionError

GetInvalidationError

GetKeyGroupConfigError

GetKeyGroupError

GetMonitoringSubscriptionError

GetOriginRequestPolicyConfigError

GetOriginRequestPolicyError

GetPublicKeyConfigError

GetPublicKeyError

GetRealtimeLogConfigError

GetStreamingDistributionConfigError

GetStreamingDistributionError

IllegalDelete

You cannot delete a managed policy.

IllegalFieldLevelEncryptionConfigAssociationWithCacheBehavior

The specified configuration for field-level encryption can't be associated with the specified cache behavior.

IllegalUpdate

The update contains modifications that are not allowed.

InconsistentQuantities

The value of Quantity and the size of Items don't match.

InvalidArgument

An argument is invalid.

InvalidDefaultRootObject

The default root object file name is too big or contains an invalid character.

InvalidErrorCode

An invalid error code was specified.

InvalidForwardCookies

Your request contains forward cookies option which doesn't match with the expectation for the whitelisted

InvalidFunctionAssociation

A CloudFront function association is invalid.

InvalidGeoRestrictionParameter

The specified geo restriction parameter is not valid.

InvalidHeadersForS3Origin

The headers specified are not valid for an Amazon S3 origin.

InvalidIfMatchVersion

The If-Match version is missing or not valid.

InvalidLambdaFunctionAssociation

The specified Lambda@Edge function association is invalid.

InvalidLocationCode

The location code specified is not valid.

InvalidMinimumProtocolVersion

The minimum protocol version specified is not valid.

InvalidOrigin

The Amazon S3 origin server specified does not refer to a valid Amazon S3 bucket.

InvalidOriginAccessIdentity

The origin access identity is not valid or doesn't exist.

InvalidOriginKeepaliveTimeout

The keep alive timeout specified for the origin is not valid.

InvalidOriginReadTimeout

The read timeout specified for the origin is not valid.

InvalidProtocolSettings

You cannot specify SSLv3 as the minimum protocol version if you only want to support only clients that support

InvalidQueryStringParameters

The query string parameters specified are not valid.

InvalidRelativePath

The relative path is too big, is not URL-encoded, or does not begin with a slash (/).

InvalidRequiredProtocol

This operation requires the HTTPS protocol. Ensure that you specify the HTTPS protocol in your request, or omit the

InvalidResponseCode

A response code is not valid.

InvalidTagging

The tagging specified is not valid.

InvalidTtlOrder

The TTL order specified is not valid.

InvalidViewerCertificate

A viewer certificate specified is not valid.

InvalidWebAclId

A web ACL ID specified is not valid. To specify a web ACL created using the latest

KeyGroupAlreadyExists

A key group with this name already exists. You must provide a unique name. To modify an

ListCachePoliciesError

ListCloudFrontOriginAccessIdentitiesError

ListConflictingAliasesError

ListDistributionsByCachePolicyIdError

ListDistributionsByKeyGroupError

ListDistributionsByOriginRequestPolicyIdError

ListDistributionsByRealtimeLogConfigError

ListDistributionsByWebACLIdError

ListDistributionsError

ListFieldLevelEncryptionConfigsError

ListFieldLevelEncryptionProfilesError

ListFunctionsError

ListInvalidationsError

ListKeyGroupsError

ListOriginRequestPoliciesError

ListPublicKeysError

ListRealtimeLogConfigsError

ListStreamingDistributionsError

ListTagsForResourceError

MissingBody

This operation requires a body. Ensure that the body is present and the Content-Type header is set.

NoSuchCachePolicy

The cache policy does not exist.

NoSuchCloudFrontOriginAccessIdentity

The specified origin access identity does not exist.

NoSuchDistribution

The specified distribution does not exist.

NoSuchFieldLevelEncryptionConfig

The specified configuration for field-level encryption doesn't exist.

NoSuchFieldLevelEncryptionProfile

The specified profile for field-level encryption doesn't exist.

NoSuchFunctionExists

The function does not exist.

NoSuchInvalidation

The specified invalidation does not exist.

NoSuchOrigin

No origin exists with the specified Origin Id.

NoSuchOriginRequestPolicy

The origin request policy does not exist.

NoSuchPublicKey

The specified public key doesn't exist.

NoSuchRealtimeLogConfig

The real-time log configuration does not exist.

NoSuchResource

A resource that was specified is not valid.

NoSuchStreamingDistribution

The specified streaming distribution does not exist.

OriginRequestPolicyAlreadyExists

An origin request policy with this name already exists. You must provide a unique

OriginRequestPolicyInUse

Cannot delete the origin request policy because it is attached to one or more cache

PreconditionFailed

The precondition in one or more of the request fields evaluated to

PublicKeyAlreadyExists

The specified public key already exists.

PublicKeyInUse

The specified public key is in use.

PublishFunctionError

QueryArgProfileEmpty

No profile specified for the field-level encryption query argument.

RealtimeLogConfigAlreadyExists

A real-time log configuration with this name already exists. You must provide a unique name.

RealtimeLogConfigInUse

Cannot delete the real-time log configuration because it is attached to one or more cache

RealtimeLogConfigOwnerMismatch

The specified real-time log configuration belongs to a different account.

ResourceInUse

Cannot delete this resource because it is in use.

StreamingDistributionAlreadyExists

The caller reference you attempted to create the streaming distribution with

StreamingDistributionNotDisabled

The specified CloudFront distribution is not disabled. You must disable

TagResourceError

TestFunctionError

TestFunctionFailed

The CloudFront function failed.

TooManyCacheBehaviors

You cannot create more cache behaviors for the distribution.

TooManyCachePolicies

You have reached the maximum number of cache policies for this account. For more

TooManyCertificates

You cannot create anymore custom SSL/TLS certificates.

TooManyCloudFrontOriginAccessIdentities

Processing your request would cause you to exceed the maximum number of origin access identities allowed.

TooManyCookieNamesInWhiteList

Your request contains more cookie names in the whitelist than are allowed per cache behavior.

TooManyCookiesInCachePolicy

The number of cookies in the cache policy exceeds the maximum. For more information,

TooManyCookiesInOriginRequestPolicy

The number of cookies in the origin request policy exceeds the maximum. For more

TooManyDistributionCnamEs

Your request contains more CNAMEs than are allowed per distribution.

TooManyDistributions

Processing your request would cause you to exceed the maximum number of distributions allowed.

TooManyDistributionsAssociatedToCachePolicy

The maximum number of distributions have been associated with the specified cache

TooManyDistributionsAssociatedToFieldLevelEncryptionConfig

The maximum number of distributions have been associated with the specified configuration for field-level encryption.

TooManyDistributionsAssociatedToKeyGroup

The number of distributions that reference this key group is more than the maximum

TooManyDistributionsAssociatedToOriginRequestPolicy

The maximum number of distributions have been associated with the specified origin

TooManyDistributionsWithFunctionAssociations

You have reached the maximum number of distributions that are associated with a CloudFront

TooManyDistributionsWithLambdaAssociations

Processing your request would cause the maximum number of distributions with Lambda@Edge function associations per owner

TooManyDistributionsWithSingleFunctionArn

The maximum number of distributions have been associated with the specified Lambda@Edge

TooManyFieldLevelEncryptionConfigs

The maximum number of configurations for field-level encryption have been created.

TooManyFieldLevelEncryptionContentTypeProfiles

The maximum number of content type profiles for field-level encryption have been created.

TooManyFieldLevelEncryptionEncryptionEntities

The maximum number of encryption entities for field-level encryption have been created.

TooManyFieldLevelEncryptionFieldPatterns

The maximum number of field patterns for field-level encryption have been created.

TooManyFieldLevelEncryptionProfiles

The maximum number of profiles for field-level encryption have been created.

TooManyFieldLevelEncryptionQueryArgProfiles

The maximum number of query arg profiles for field-level encryption have been created.

TooManyFunctionAssociations

You have reached the maximum number of CloudFront function associations for this

TooManyFunctions

You have reached the maximum number of CloudFront functions for this account. For more

TooManyHeadersInCachePolicy

The number of headers in the cache policy exceeds the maximum. For more information,

TooManyHeadersInForwardedValues

Your request contains too many headers in forwarded values.

TooManyHeadersInOriginRequestPolicy

The number of headers in the origin request policy exceeds the maximum. For more

TooManyInvalidationsInProgress

You have exceeded the maximum number of allowable InProgress invalidation batch requests, or invalidation objects.

TooManyKeyGroups

You have reached the maximum number of key groups for this account. For more

TooManyKeyGroupsAssociatedToDistribution

The number of key groups referenced by this distribution is more than the maximum

TooManyLambdaFunctionAssociations

Your request contains more Lambda@Edge function associations than are allowed per distribution.

TooManyOriginCustomHeaders

Your request contains too many origin custom headers.

TooManyOriginGroupsPerDistribution

Processing your request would cause you to exceed the maximum number of origin groups allowed.

TooManyOriginRequestPolicies

You have reached the maximum number of origin request policies for this account.

TooManyOrigins

You cannot create more origins for the distribution.

TooManyPublicKeys

The maximum number of public keys for field-level encryption have been created. To create a new public key, delete one of the existing keys.

TooManyPublicKeysInKeyGroup

The number of public keys in this key group is more than the maximum allowed. For more

TooManyQueryStringParameters

Your request contains too many query string parameters.

TooManyQueryStringsInCachePolicy

The number of query strings in the cache policy exceeds the maximum. For more

TooManyQueryStringsInOriginRequestPolicy

The number of query strings in the origin request policy exceeds the maximum. For more

TooManyRealtimeLogConfigs

You have reached the maximum number of real-time log configurations for this account.

TooManyStreamingDistributionCnamEs

Your request contains more CNAMEs than are allowed per distribution.

TooManyStreamingDistributions

Processing your request would cause you to exceed the maximum number of streaming distributions allowed.

TooManyTrustedSigners

Your request contains more trusted signers than are allowed per distribution.

TrustedKeyGroupDoesNotExist

The specified key group does not exist.

TrustedSignerDoesNotExist

One or more of your trusted signers don't exist.

UnsupportedOperation

This operation is not supported in this region.

UntagResourceError

UpdateCachePolicyError

UpdateCloudFrontOriginAccessIdentityError

UpdateDistributionError

UpdateFieldLevelEncryptionConfigError

UpdateFieldLevelEncryptionProfileError

UpdateFunctionError

UpdateKeyGroupError

UpdateOriginRequestPolicyError

UpdatePublicKeyError

UpdateRealtimeLogConfigError

UpdateStreamingDistributionError

Enums

AssociateAliasErrorKind

CreateCachePolicyErrorKind

CreateCloudFrontOriginAccessIdentityErrorKind

CreateDistributionErrorKind

CreateDistributionWithTagsErrorKind

CreateFieldLevelEncryptionConfigErrorKind

CreateFieldLevelEncryptionProfileErrorKind

CreateFunctionErrorKind

CreateInvalidationErrorKind

CreateKeyGroupErrorKind

CreateMonitoringSubscriptionErrorKind

CreateOriginRequestPolicyErrorKind

CreatePublicKeyErrorKind

CreateRealtimeLogConfigErrorKind

CreateStreamingDistributionErrorKind

CreateStreamingDistributionWithTagsErrorKind

DeleteCachePolicyErrorKind

DeleteCloudFrontOriginAccessIdentityErrorKind

DeleteDistributionErrorKind

DeleteFieldLevelEncryptionConfigErrorKind

DeleteFieldLevelEncryptionProfileErrorKind

DeleteFunctionErrorKind

DeleteKeyGroupErrorKind

DeleteMonitoringSubscriptionErrorKind

DeleteOriginRequestPolicyErrorKind

DeletePublicKeyErrorKind

DeleteRealtimeLogConfigErrorKind

DeleteStreamingDistributionErrorKind

DescribeFunctionErrorKind

GetCachePolicyConfigErrorKind

GetCachePolicyErrorKind

GetCloudFrontOriginAccessIdentityConfigErrorKind

GetCloudFrontOriginAccessIdentityErrorKind

GetDistributionConfigErrorKind

GetDistributionErrorKind

GetFieldLevelEncryptionConfigErrorKind

GetFieldLevelEncryptionErrorKind

GetFieldLevelEncryptionProfileConfigErrorKind

GetFieldLevelEncryptionProfileErrorKind

GetFunctionErrorKind

GetInvalidationErrorKind

GetKeyGroupConfigErrorKind

GetKeyGroupErrorKind

GetMonitoringSubscriptionErrorKind

GetOriginRequestPolicyConfigErrorKind

GetOriginRequestPolicyErrorKind

GetPublicKeyConfigErrorKind

GetPublicKeyErrorKind

GetRealtimeLogConfigErrorKind

GetStreamingDistributionConfigErrorKind

GetStreamingDistributionErrorKind

ListCachePoliciesErrorKind

ListCloudFrontOriginAccessIdentitiesErrorKind

ListConflictingAliasesErrorKind

ListDistributionsByCachePolicyIdErrorKind

ListDistributionsByKeyGroupErrorKind

ListDistributionsByOriginRequestPolicyIdErrorKind

ListDistributionsByRealtimeLogConfigErrorKind

ListDistributionsByWebACLIdErrorKind

ListDistributionsErrorKind

ListFieldLevelEncryptionConfigsErrorKind

ListFieldLevelEncryptionProfilesErrorKind

ListFunctionsErrorKind

ListInvalidationsErrorKind

ListKeyGroupsErrorKind

ListOriginRequestPoliciesErrorKind

ListPublicKeysErrorKind

ListRealtimeLogConfigsErrorKind

ListStreamingDistributionsErrorKind

ListTagsForResourceErrorKind

PublishFunctionErrorKind

TagResourceErrorKind

TestFunctionErrorKind

UntagResourceErrorKind

UpdateCachePolicyErrorKind

UpdateCloudFrontOriginAccessIdentityErrorKind

UpdateDistributionErrorKind

UpdateFieldLevelEncryptionConfigErrorKind

UpdateFieldLevelEncryptionProfileErrorKind

UpdateFunctionErrorKind

UpdateKeyGroupErrorKind

UpdateOriginRequestPolicyErrorKind

UpdatePublicKeyErrorKind

UpdateRealtimeLogConfigErrorKind

UpdateStreamingDistributionErrorKind