Module aws_sdk_cloudfront::input

Modules

associate_alias_input

See AssociateAliasInput

create_cache_policy_input

See CreateCachePolicyInput

create_cloud_front_origin_access_identity_input

See CreateCloudFrontOriginAccessIdentityInput

create_distribution_input

See CreateDistributionInput

create_distribution_with_tags_input

See CreateDistributionWithTagsInput

create_field_level_encryption_config_input

See CreateFieldLevelEncryptionConfigInput

create_field_level_encryption_profile_input

See CreateFieldLevelEncryptionProfileInput

create_function_input

See CreateFunctionInput

create_invalidation_input

See CreateInvalidationInput

create_key_group_input

See CreateKeyGroupInput

create_monitoring_subscription_input

See CreateMonitoringSubscriptionInput

create_origin_request_policy_input

See CreateOriginRequestPolicyInput

create_public_key_input

See CreatePublicKeyInput

create_realtime_log_config_input

See CreateRealtimeLogConfigInput

create_streaming_distribution_input

See CreateStreamingDistributionInput

create_streaming_distribution_with_tags_input

See CreateStreamingDistributionWithTagsInput

delete_cache_policy_input

See DeleteCachePolicyInput

delete_cloud_front_origin_access_identity_input

See DeleteCloudFrontOriginAccessIdentityInput

delete_distribution_input

See DeleteDistributionInput

delete_field_level_encryption_config_input

See DeleteFieldLevelEncryptionConfigInput

delete_field_level_encryption_profile_input

See DeleteFieldLevelEncryptionProfileInput

delete_function_input

See DeleteFunctionInput

delete_key_group_input

See DeleteKeyGroupInput

delete_monitoring_subscription_input

See DeleteMonitoringSubscriptionInput

delete_origin_request_policy_input

See DeleteOriginRequestPolicyInput

delete_public_key_input

See DeletePublicKeyInput

delete_realtime_log_config_input

See DeleteRealtimeLogConfigInput

delete_streaming_distribution_input

See DeleteStreamingDistributionInput

describe_function_input

See DescribeFunctionInput

get_cache_policy_config_input

See GetCachePolicyConfigInput

get_cache_policy_input

See GetCachePolicyInput

get_cloud_front_origin_access_identity_config_input

See GetCloudFrontOriginAccessIdentityConfigInput

get_cloud_front_origin_access_identity_input

See GetCloudFrontOriginAccessIdentityInput

get_distribution_config_input

See GetDistributionConfigInput

get_distribution_input

See GetDistributionInput

get_field_level_encryption_config_input

See GetFieldLevelEncryptionConfigInput

get_field_level_encryption_input

See GetFieldLevelEncryptionInput

get_field_level_encryption_profile_config_input

See GetFieldLevelEncryptionProfileConfigInput

get_field_level_encryption_profile_input

See GetFieldLevelEncryptionProfileInput

get_function_input

See GetFunctionInput

get_invalidation_input

See GetInvalidationInput

get_key_group_config_input

See GetKeyGroupConfigInput

get_key_group_input

See GetKeyGroupInput

get_monitoring_subscription_input

See GetMonitoringSubscriptionInput

get_origin_request_policy_config_input

See GetOriginRequestPolicyConfigInput

get_origin_request_policy_input

See GetOriginRequestPolicyInput

get_public_key_config_input

See GetPublicKeyConfigInput

get_public_key_input

See GetPublicKeyInput

get_realtime_log_config_input

See GetRealtimeLogConfigInput

get_streaming_distribution_config_input

See GetStreamingDistributionConfigInput

get_streaming_distribution_input

See GetStreamingDistributionInput

list_cache_policies_input

See ListCachePoliciesInput

list_cloud_front_origin_access_identities_input

See ListCloudFrontOriginAccessIdentitiesInput

list_conflicting_aliases_input

See ListConflictingAliasesInput

list_distributions_by_cache_policy_id_input

See ListDistributionsByCachePolicyIdInput

list_distributions_by_key_group_input

See ListDistributionsByKeyGroupInput

list_distributions_by_origin_request_policy_id_input

See ListDistributionsByOriginRequestPolicyIdInput

list_distributions_by_realtime_log_config_input

See ListDistributionsByRealtimeLogConfigInput

list_distributions_by_web_acl_id_input

See ListDistributionsByWebAclIdInput

list_distributions_input

See ListDistributionsInput

list_field_level_encryption_configs_input

See ListFieldLevelEncryptionConfigsInput

list_field_level_encryption_profiles_input

See ListFieldLevelEncryptionProfilesInput

list_functions_input

See ListFunctionsInput

list_invalidations_input

See ListInvalidationsInput

list_key_groups_input

See ListKeyGroupsInput

list_origin_request_policies_input

See ListOriginRequestPoliciesInput

list_public_keys_input

See ListPublicKeysInput

list_realtime_log_configs_input

See ListRealtimeLogConfigsInput

list_streaming_distributions_input

See ListStreamingDistributionsInput

list_tags_for_resource_input

See ListTagsForResourceInput

publish_function_input

See PublishFunctionInput

tag_resource_input

See TagResourceInput

test_function_input

See TestFunctionInput

untag_resource_input

See UntagResourceInput

update_cache_policy_input

See UpdateCachePolicyInput

update_cloud_front_origin_access_identity_input

See UpdateCloudFrontOriginAccessIdentityInput

update_distribution_input

See UpdateDistributionInput

update_field_level_encryption_config_input

See UpdateFieldLevelEncryptionConfigInput

update_field_level_encryption_profile_input

See UpdateFieldLevelEncryptionProfileInput

update_function_input

See UpdateFunctionInput

update_key_group_input

See UpdateKeyGroupInput

update_origin_request_policy_input

See UpdateOriginRequestPolicyInput

update_public_key_input

See UpdatePublicKeyInput

update_realtime_log_config_input

See UpdateRealtimeLogConfigInput

update_streaming_distribution_input

See UpdateStreamingDistributionInput

Structs

AssociateAliasInput

CreateCachePolicyInput

CreateCloudFrontOriginAccessIdentityInput

The request to create a new origin access identity (OAI). An origin access identity is a

CreateDistributionInput

The request to create a new distribution.

CreateDistributionWithTagsInput

The request to create a new distribution with tags.

CreateFieldLevelEncryptionConfigInput

CreateFieldLevelEncryptionProfileInput

CreateFunctionInput

CreateInvalidationInput

The request to create an invalidation.

CreateKeyGroupInput

CreateMonitoringSubscriptionInput

CreateOriginRequestPolicyInput

CreatePublicKeyInput

CreateRealtimeLogConfigInput

CreateStreamingDistributionInput

The request to create a new streaming distribution.

CreateStreamingDistributionWithTagsInput

The request to create a new streaming distribution with tags.

DeleteCachePolicyInput

DeleteCloudFrontOriginAccessIdentityInput

Deletes a origin access identity.

DeleteDistributionInput

This action deletes a web distribution. To delete a web distribution using the CloudFront

DeleteFieldLevelEncryptionConfigInput

DeleteFieldLevelEncryptionProfileInput

DeleteFunctionInput

DeleteKeyGroupInput

DeleteMonitoringSubscriptionInput

DeleteOriginRequestPolicyInput

DeletePublicKeyInput

DeleteRealtimeLogConfigInput

DeleteStreamingDistributionInput

The request to delete a streaming distribution.

DescribeFunctionInput

GetCachePolicyConfigInput

GetCachePolicyInput

GetCloudFrontOriginAccessIdentityConfigInput

The origin access identity's configuration information. For more information, see

GetCloudFrontOriginAccessIdentityInput

The request to get an origin access identity's information.

GetDistributionConfigInput

The request to get a distribution configuration.

GetDistributionInput

The request to get a distribution's information.

GetFieldLevelEncryptionConfigInput

GetFieldLevelEncryptionInput

GetFieldLevelEncryptionProfileConfigInput

GetFieldLevelEncryptionProfileInput

GetFunctionInput

GetInvalidationInput

The request to get an invalidation's information.

GetKeyGroupConfigInput

GetKeyGroupInput

GetMonitoringSubscriptionInput

GetOriginRequestPolicyConfigInput

GetOriginRequestPolicyInput

GetPublicKeyConfigInput

GetPublicKeyInput

GetRealtimeLogConfigInput

GetStreamingDistributionConfigInput

To request to get a streaming distribution configuration.

GetStreamingDistributionInput

The request to get a streaming distribution's information.

ListCachePoliciesInput

ListCloudFrontOriginAccessIdentitiesInput

The request to list origin access identities.

ListConflictingAliasesInput

ListDistributionsByCachePolicyIdInput

ListDistributionsByKeyGroupInput

ListDistributionsByOriginRequestPolicyIdInput

ListDistributionsByRealtimeLogConfigInput

ListDistributionsByWebAclIdInput

The request to list distributions that are associated with a specified WAF web

ListDistributionsInput

The request to list your distributions.

ListFieldLevelEncryptionConfigsInput

ListFieldLevelEncryptionProfilesInput

ListFunctionsInput

ListInvalidationsInput

The request to list invalidations.

ListKeyGroupsInput

ListOriginRequestPoliciesInput

ListPublicKeysInput

ListRealtimeLogConfigsInput

ListStreamingDistributionsInput

The request to list your streaming distributions.

ListTagsForResourceInput

The request to list tags for a CloudFront resource.

PublishFunctionInput

TagResourceInput

The request to add tags to a CloudFront resource.

TestFunctionInput

UntagResourceInput

The request to remove tags from a CloudFront resource.

UpdateCachePolicyInput

UpdateCloudFrontOriginAccessIdentityInput

The request to update an origin access identity.

UpdateDistributionInput

The request to update a distribution.

UpdateFieldLevelEncryptionConfigInput

UpdateFieldLevelEncryptionProfileInput

UpdateFunctionInput

UpdateKeyGroupInput

UpdateOriginRequestPolicyInput

UpdatePublicKeyInput

UpdateRealtimeLogConfigInput

UpdateStreamingDistributionInput

The request to update a streaming distribution.