Module aws_sdk_cloudfront::model

Modules

active_trusted_key_groups

See ActiveTrustedKeyGroups

active_trusted_signers

See ActiveTrustedSigners

alias_icp_recordal

See AliasIcpRecordal

aliases

See Aliases

allowed_methods

See AllowedMethods

cache_behavior

See CacheBehavior

cache_behaviors

See CacheBehaviors

cache_policy

See CachePolicy

cache_policy_config

See CachePolicyConfig

cache_policy_cookies_config

See CachePolicyCookiesConfig

cache_policy_headers_config

See CachePolicyHeadersConfig

cache_policy_list

See CachePolicyList

cache_policy_query_strings_config

See CachePolicyQueryStringsConfig

cache_policy_summary

See CachePolicySummary

cached_methods

See CachedMethods

cloud_front_origin_access_identity

See CloudFrontOriginAccessIdentity

cloud_front_origin_access_identity_config

See CloudFrontOriginAccessIdentityConfig

cloud_front_origin_access_identity_list

See CloudFrontOriginAccessIdentityList

cloud_front_origin_access_identity_summary

See CloudFrontOriginAccessIdentitySummary

conflicting_alias

See ConflictingAlias

conflicting_aliases_list

See ConflictingAliasesList

content_type_profile

See ContentTypeProfile

content_type_profile_config

See ContentTypeProfileConfig

content_type_profiles

See ContentTypeProfiles

cookie_names

See CookieNames

cookie_preference

See CookiePreference

custom_error_response

See CustomErrorResponse

custom_error_responses

See CustomErrorResponses

custom_headers

See CustomHeaders

custom_origin_config

See CustomOriginConfig

default_cache_behavior

See DefaultCacheBehavior

distribution

See Distribution

distribution_config

See DistributionConfig

distribution_config_with_tags

See DistributionConfigWithTags

distribution_id_list

See DistributionIdList

distribution_list

See DistributionList

distribution_summary

See DistributionSummary

encryption_entities

See EncryptionEntities

encryption_entity

See EncryptionEntity

end_point

See EndPoint

field_level_encryption

See FieldLevelEncryption

field_level_encryption_config

See FieldLevelEncryptionConfig

field_level_encryption_list

See FieldLevelEncryptionList

field_level_encryption_profile

See FieldLevelEncryptionProfile

field_level_encryption_profile_config

See FieldLevelEncryptionProfileConfig

field_level_encryption_profile_list

See FieldLevelEncryptionProfileList

field_level_encryption_profile_summary

See FieldLevelEncryptionProfileSummary

field_level_encryption_summary

See FieldLevelEncryptionSummary

field_patterns

See FieldPatterns

forwarded_values

See ForwardedValues

function_association

See FunctionAssociation

function_associations

See FunctionAssociations

function_config

See FunctionConfig

function_list

See FunctionList

function_metadata

See FunctionMetadata

function_summary

See FunctionSummary

geo_restriction

See GeoRestriction

headers

See Headers

invalidation

See Invalidation

invalidation_batch

See InvalidationBatch

invalidation_list

See InvalidationList

invalidation_summary

See InvalidationSummary

key_group

See KeyGroup

key_group_config

See KeyGroupConfig

key_group_list

See KeyGroupList

key_group_summary

See KeyGroupSummary

key_pair_ids

See KeyPairIds

kg_key_pair_ids

See KgKeyPairIds

kinesis_stream_config

See KinesisStreamConfig

lambda_function_association

See LambdaFunctionAssociation

lambda_function_associations

See LambdaFunctionAssociations

logging_config

See LoggingConfig

monitoring_subscription

See MonitoringSubscription

origin

See Origin

origin_custom_header

See OriginCustomHeader

origin_group

See OriginGroup

origin_group_failover_criteria

See OriginGroupFailoverCriteria

origin_group_member

See OriginGroupMember

origin_group_members

See OriginGroupMembers

origin_groups

See OriginGroups

origin_request_policy

See OriginRequestPolicy

origin_request_policy_config

See OriginRequestPolicyConfig

origin_request_policy_cookies_config

See OriginRequestPolicyCookiesConfig

origin_request_policy_headers_config

See OriginRequestPolicyHeadersConfig

origin_request_policy_list

See OriginRequestPolicyList

origin_request_policy_query_strings_config

See OriginRequestPolicyQueryStringsConfig

origin_request_policy_summary

See OriginRequestPolicySummary

origin_shield

See OriginShield

origin_ssl_protocols

See OriginSslProtocols

origins

See Origins

parameters_in_cache_key_and_forwarded_to_origin

See ParametersInCacheKeyAndForwardedToOrigin

paths

See Paths

public_key

See PublicKey

public_key_config

See PublicKeyConfig

public_key_list

See PublicKeyList

public_key_summary

See PublicKeySummary

query_arg_profile

See QueryArgProfile

query_arg_profile_config

See QueryArgProfileConfig

query_arg_profiles

See QueryArgProfiles

query_string_cache_keys

See QueryStringCacheKeys

query_string_names

See QueryStringNames

realtime_log_config

See RealtimeLogConfig

realtime_log_configs

See RealtimeLogConfigs

realtime_metrics_subscription_config

See RealtimeMetricsSubscriptionConfig

restrictions

See Restrictions

s3_origin

See S3Origin

s3_origin_config

See S3OriginConfig

signer

See Signer

status_codes

See StatusCodes

streaming_distribution

See StreamingDistribution

streaming_distribution_config

See StreamingDistributionConfig

streaming_distribution_config_with_tags

See StreamingDistributionConfigWithTags

streaming_distribution_list

See StreamingDistributionList

streaming_distribution_summary

See StreamingDistributionSummary

streaming_logging_config

See StreamingLoggingConfig

tag

See Tag

tag_keys

See TagKeys

tags

See Tags

test_result

See TestResult

trusted_key_groups

See TrustedKeyGroups

trusted_signers

See TrustedSigners

viewer_certificate

See ViewerCertificate

Structs

ActiveTrustedKeyGroups

A list of key groups, and the public keys in each key group,

ActiveTrustedSigners

A list of accounts and the active CloudFront key pairs in each account that CloudFront can use

AliasIcpRecordal

Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content

Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any,

AllowedMethods

A complex type that controls which HTTP methods CloudFront processes and forwards to your

CacheBehavior

A complex type that describes how CloudFront processes requests.

CacheBehaviors

A complex type that contains zero or more CacheBehavior elements.

CachePolicy

A cache policy.

CachePolicyConfig

A cache policy configuration.

CachePolicyCookiesConfig

An object that determines whether any cookies in viewer requests (and if so, which cookies)

CachePolicyHeadersConfig

An object that determines whether any HTTP headers (and if so, which headers) are included

CachePolicyList

A list of cache policies.

CachePolicyQueryStringsConfig

An object that determines whether any URL query strings in viewer requests (and if so, which

CachePolicySummary

Contains a cache policy.

CachedMethods

A complex type that controls whether CloudFront caches the response to requests using the

CloudFrontOriginAccessIdentity

CloudFront origin access identity.

CloudFrontOriginAccessIdentityConfig

Origin access identity configuration. Send a GET request to the

CloudFrontOriginAccessIdentityList

Lists the origin access identities for CloudFront.Send a GET request to the

CloudFrontOriginAccessIdentitySummary

Summary of the information about a CloudFront origin access identity.

ConflictingAlias

An alias (also called a CNAME) and the CloudFront distribution and Amazon Web Services account ID that it’s

ConflictingAliasesList

A list of aliases (also called CNAMEs) and the CloudFront distributions and Amazon Web Services accounts that

ContentTypeProfile

A field-level encryption content type profile.

ContentTypeProfileConfig

The configuration for a field-level encryption content type-profile mapping.

ContentTypeProfiles

Field-level encryption content type-profile.

CookieNames

Contains a list of cookie names.

CookiePreference

This field is deprecated. We recommend that you use a cache policy or an origin

CustomErrorResponse

A complex type that controls:

CustomErrorResponses

A complex type that controls:

CustomHeaders

A complex type that contains the list of Custom Headers for each origin.

CustomOriginConfig

A custom origin. A custom origin is any origin that is not an Amazon S3

DefaultCacheBehavior

A complex type that describes the default cache behavior if you don’t specify a

Distribution

A distribution tells CloudFront where you want content to be delivered from, and the details about how to

DistributionConfig

A distribution configuration.

DistributionConfigWithTags

A distribution Configuration and a list of tags to be associated with the

DistributionIdList

A list of distribution IDs.

DistributionList

A distribution list.

DistributionSummary

A summary of the information about a CloudFront distribution.

EncryptionEntities

Complex data type for field-level encryption profiles that includes all of the encryption entities.

EncryptionEntity

Complex data type for field-level encryption profiles that includes the encryption key and field pattern specifications.

EndPoint

Contains information about the Amazon Kinesis data stream where you are sending real-time

FieldLevelEncryption

A complex data type that includes the profile configurations and other options specified for field-level encryption.

FieldLevelEncryptionConfig

A complex data type that includes the profile configurations specified for field-level encryption.

FieldLevelEncryptionList

List of field-level encrpytion configurations.

FieldLevelEncryptionProfile

A complex data type for field-level encryption profiles.

FieldLevelEncryptionProfileConfig

A complex data type of profiles for the field-level encryption.

FieldLevelEncryptionProfileList

List of field-level encryption profiles.

FieldLevelEncryptionProfileSummary

The field-level encryption profile summary.

FieldLevelEncryptionSummary

A summary of a field-level encryption item.

FieldPatterns

A complex data type that includes the field patterns to match for field-level encryption.

ForwardedValues

This field is deprecated. We recommend that you use a cache policy or an origin

FunctionAssociation

A CloudFront function that is associated with a cache behavior in a CloudFront

FunctionAssociations

A list of CloudFront functions that are associated with a cache behavior in a CloudFront distribution.

FunctionConfig

Contains configuration information about a CloudFront function.

FunctionList

A list of CloudFront functions.

FunctionMetadata

Contains metadata about a CloudFront function.

FunctionSummary

Contains configuration information and metadata about a CloudFront function.

GeoRestriction

A complex type that controls the countries in which your content is distributed. CloudFront

Headers

Contains a list of HTTP header names.

Invalidation

An invalidation.

InvalidationBatch

An invalidation batch.

InvalidationList

The InvalidationList complex type describes the list of invalidation

InvalidationSummary

A summary of an invalidation request.

KeyGroup

A key group.

KeyGroupConfig

A key group configuration.

KeyGroupList

A list of key groups.

KeyGroupSummary

Contains information about a key group.

KeyPairIds

A list of CloudFront key pair identifiers.

KgKeyPairIds

A list of identifiers for the public keys that CloudFront can use to verify the

KinesisStreamConfig

Contains information about the Amazon Kinesis data stream where you are sending

LambdaFunctionAssociation

A complex type that contains a Lambda@Edge function association.

LambdaFunctionAssociations

A complex type that specifies a list of Lambda@Edge functions associations for a cache

LoggingConfig

A complex type that controls whether access logs are written for the

MonitoringSubscription

A monitoring subscription. This structure contains information about whether additional

Origin

An origin.

OriginCustomHeader

A complex type that contains HeaderName and HeaderValue

OriginGroup

An origin group includes two origins (a primary origin and a second origin to failover to) and a failover criteria that you specify.

OriginGroupFailoverCriteria

A complex data type that includes information about the failover criteria for an origin group, including

OriginGroupMember

An origin in an origin group.

OriginGroupMembers

A complex data type for the origins included in an origin group.

OriginGroups

A complex data type for the origin groups specified for a distribution.

OriginRequestPolicy

An origin request policy.

OriginRequestPolicyConfig

An origin request policy configuration.

OriginRequestPolicyCookiesConfig

An object that determines whether any cookies in viewer requests (and if so, which cookies)

OriginRequestPolicyHeadersConfig

An object that determines whether any HTTP headers (and if so, which headers) are included

OriginRequestPolicyList

A list of origin request policies.

OriginRequestPolicyQueryStringsConfig

An object that determines whether any URL query strings in viewer requests (and if so, which

OriginRequestPolicySummary

Contains an origin request policy.

OriginShield

CloudFront Origin Shield.

OriginSslProtocols

A complex type that contains information about the SSL/TLS protocols that CloudFront can use

Origins

Contains information about the origins for this distribution.

ParametersInCacheKeyAndForwardedToOrigin

This object determines the values that CloudFront includes in the cache key. These values can

Paths

A complex type that contains information about the objects that you want to invalidate.

PublicKey

A public key that you can use with signed URLs and signed cookies, or with field-level encryption.

PublicKeyConfig

Configuration information about a public key that you can use with signed URLs and signed cookies, or with field-level encryption.

PublicKeyList

A list of public keys that you can use with signed URLs and signed cookies, or with field-level encryption.

PublicKeySummary

Contains information about a public key.

QueryArgProfile

Query argument-profile mapping for field-level encryption.

QueryArgProfileConfig

Configuration for query argument-profile mapping for field-level encryption.

QueryArgProfiles

Query argument-profile mapping for field-level encryption.

QueryStringCacheKeys

This field is deprecated. We recommend that you use a cache policy or an origin

QueryStringNames

Contains a list of query string names.

RealtimeLogConfig

A real-time log configuration.

RealtimeLogConfigs

A list of real-time log configurations.

RealtimeMetricsSubscriptionConfig

A subscription configuration for additional CloudWatch metrics.

Restrictions

A complex type that identifies ways in which you want to restrict distribution of your

S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want

S3OriginConfig

A complex type that contains information about the Amazon S3 origin. If the origin is a

Signer

A list of accounts and the active CloudFront key pairs in each account that CloudFront can use to

StatusCodes

A complex data type for the status codes that you specify that, when returned by a primary origin, trigger

StreamingDistribution

A streaming distribution tells CloudFront where you want RTMP content to be delivered from, and the details about how to

StreamingDistributionConfig

The RTMP distribution's configuration information.

StreamingDistributionConfigWithTags

A streaming distribution Configuration and a list of tags to be associated with the

StreamingDistributionList

A streaming distribution list.

StreamingDistributionSummary

A summary of the information for a CloudFront streaming distribution.

StreamingLoggingConfig

A complex type that controls whether access logs are written for this streaming distribution.

Tag

A complex type that contains Tag key and Tag value.

TagKeys

A complex type that contains zero or more Tag elements.

Tags

A complex type that contains zero or more Tag elements.

TestResult

Contains the result of testing a CloudFront function with TestFunction.

TrustedKeyGroups

A list of key groups whose public keys CloudFront can use to verify the signatures of signed

TrustedSigners

A list of accounts whose public keys CloudFront can use to verify the signatures of signed

ViewerCertificate

A complex type that determines the distribution’s SSL/TLS configuration for communicating

Enums

CachePolicyCookieBehavior

CachePolicyHeaderBehavior

CachePolicyQueryStringBehavior

CachePolicyType

CertificateSource

EventType

Format

FunctionRuntime

FunctionStage

GeoRestrictionType

HttpVersion

IcpRecordalStatus

ItemSelection

Method

MinimumProtocolVersion

OriginProtocolPolicy

OriginRequestPolicyCookieBehavior

OriginRequestPolicyHeaderBehavior

OriginRequestPolicyQueryStringBehavior

OriginRequestPolicyType

PriceClass

RealtimeMetricsSubscriptionStatus

SslProtocol

SslSupportMethod

ViewerProtocolPolicy