Struct aws_sdk_cloudfront::model::default_cache_behavior::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DefaultCacheBehavior

Implementations

impl Builder

pub fn target_origin_id(self, input: impl Into<String>) -> Self

The value of ID for the origin that you want CloudFront to route requests to when they use the default cache behavior.

pub fn set_target_origin_id(self, input: Option<String>) -> Self

pub fn trusted_signers(self, input: TrustedSigners) -> Self

We recommend using TrustedKeyGroups instead of TrustedSigners.

A list of account IDs whose public keys CloudFront can use to validate signed URLs or signed cookies.

When a cache behavior contains trusted signers, CloudFront requires signed URLs or signed cookies for all requests that match the cache behavior. The URLs or cookies must be signed with the private key of a CloudFront key pair in a trusted signer’s account. The signed URL or cookie contains information about which public key CloudFront should use to verify the signature. For more information, see Serving private content in the Amazon CloudFront Developer Guide.

pub fn set_trusted_signers(self, input: Option<TrustedSigners>) -> Self

pub fn trusted_key_groups(self, input: TrustedKeyGroups) -> Self

A list of key groups that CloudFront can use to validate signed URLs or signed cookies.

When a cache behavior contains trusted key groups, CloudFront requires signed URLs or signed cookies for all requests that match the cache behavior. The URLs or cookies must be signed with a private key whose corresponding public key is in the key group. The signed URL or cookie contains information about which public key CloudFront should use to verify the signature. For more information, see Serving private content in the Amazon CloudFront Developer Guide.

pub fn set_trusted_key_groups(self, input: Option<TrustedKeyGroups>) -> Self

pub fn viewer_protocol_policy(self, input: ViewerProtocolPolicy) -> Self

The protocol that viewers can use to access the files in the origin specified by TargetOriginId when a request matches the path pattern in PathPattern. You can specify the following options:

allow-all: Viewers can use HTTP or HTTPS.

redirect-to-https: If a viewer submits an HTTP request, CloudFront returns an HTTP status code of 301 (Moved Permanently) to the viewer along with the HTTPS URL. The viewer then resubmits the request using the new URL.

https-only: If a viewer sends an HTTP request, CloudFront returns an HTTP status code of 403 (Forbidden).

For more information about requiring the HTTPS protocol, see Requiring HTTPS Between Viewers and CloudFront in the Amazon CloudFront Developer Guide.

The only way to guarantee that viewers retrieve an object that was fetched from the origin using HTTPS is never to use any other protocol to fetch the object. If you have recently changed from HTTP to HTTPS, we recommend that you clear your objects’ cache because cached objects are protocol agnostic. That means that an edge location will return an object from the cache regardless of whether the current request protocol matches the protocol used previously. For more information, see Managing Cache Expiration in the Amazon CloudFront Developer Guide.

pub fn set_viewer_protocol_policy(
self,
input: Option<ViewerProtocolPolicy>
) -> Self

pub fn allowed_methods(self, input: AllowedMethods) -> Self

A complex type that controls which HTTP methods CloudFront processes and forwards to your Amazon S3 bucket or your custom origin. There are three choices:

CloudFront forwards only GET and HEAD requests.

CloudFront forwards only GET, HEAD, and OPTIONS requests.

CloudFront forwards GET, HEAD, OPTIONS, PUT, PATCH, POST, and DELETE requests.

If you pick the third choice, you may need to restrict access to your Amazon S3 bucket or to your custom origin so users can't perform operations that you don't want them to. For example, you might not want users to have permissions to delete objects from your origin.

pub fn set_allowed_methods(self, input: Option<AllowedMethods>) -> Self

pub fn smooth_streaming(self, input: bool) -> Self

Indicates whether you want to distribute media files in the Microsoft Smooth Streaming format using the origin that is associated with this cache behavior. If so, specify true; if not, specify false. If you specify true for SmoothStreaming, you can still distribute other content using this cache behavior if the content matches the value of PathPattern.

pub fn set_smooth_streaming(self, input: Option<bool>) -> Self

pub fn compress(self, input: bool) -> Self

Whether you want CloudFront to automatically compress certain files for this cache behavior. If so, specify true; if not, specify false. For more information, see Serving Compressed Files in the Amazon CloudFront Developer Guide.

pub fn set_compress(self, input: Option<bool>) -> Self

pub fn lambda_function_associations(
self,
input: LambdaFunctionAssociations
) -> Self

A complex type that contains zero or more Lambda@Edge function associations for a cache behavior.

pub fn set_lambda_function_associations(
self,
input: Option<LambdaFunctionAssociations>
) -> Self

pub fn function_associations(self, input: FunctionAssociations) -> Self

A list of CloudFront functions that are associated with this cache behavior. CloudFront functions must be published to the LIVE stage to associate them with a cache behavior.

pub fn set_function_associations(
self,
input: Option<FunctionAssociations>
) -> Self

pub fn field_level_encryption_id(self, input: impl Into<String>) -> Self

The value of ID for the field-level encryption configuration that you want CloudFront to use for encrypting specific fields of data for the default cache behavior.

pub fn set_field_level_encryption_id(self, input: Option<String>) -> Self

pub fn realtime_log_config_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the real-time log configuration that is attached to this cache behavior. For more information, see Real-time logs in the Amazon CloudFront Developer Guide.

pub fn set_realtime_log_config_arn(self, input: Option<String>) -> Self

pub fn cache_policy_id(self, input: impl Into<String>) -> Self

The unique identifier of the cache policy that is attached to the default cache behavior. For more information, see Creating cache policies or Using the managed cache policies in the Amazon CloudFront Developer Guide.

A DefaultCacheBehavior must include either a CachePolicyId or ForwardedValues. We recommend that you use a CachePolicyId.

pub fn set_cache_policy_id(self, input: Option<String>) -> Self

pub fn origin_request_policy_id(self, input: impl Into<String>) -> Self

The unique identifier of the origin request policy that is attached to the default cache behavior. For more information, see Creating origin request policies or Using the managed origin request policies in the Amazon CloudFront Developer Guide.

pub fn set_origin_request_policy_id(self, input: Option<String>) -> Self

pub fn forwarded_values(self, input: ForwardedValues) -> Self

This field is deprecated. We recommend that you use a cache policy or an origin request policy instead of this field. For more information, see Working with policies in the Amazon CloudFront Developer Guide.

If you want to include values in the cache key, use a cache policy. For more information, see Creating cache policies or Using the managed cache policies in the Amazon CloudFront Developer Guide.

If you want to send values to the origin but not include them in the cache key, use an origin request policy. For more information, see Creating origin request policies or Using the managed origin request policies in the Amazon CloudFront Developer Guide.

A DefaultCacheBehavior must include either a CachePolicyId or ForwardedValues. We recommend that you use a CachePolicyId.

A complex type that specifies how CloudFront handles query strings, cookies, and HTTP headers.

pub fn set_forwarded_values(self, input: Option<ForwardedValues>) -> Self

pub fn min_ttl(self, input: i64) -> Self

This field is deprecated. We recommend that you use the MinTTL field in a cache policy instead of this field. For more information, see Creating cache policies or Using the managed cache policies in the Amazon CloudFront Developer Guide.

The minimum amount of time that you want objects to stay in CloudFront caches before CloudFront forwards another request to your origin to determine whether the object has been updated. For more information, see Managing How Long Content Stays in an Edge Cache (Expiration) in the Amazon CloudFront Developer Guide.

You must specify 0 for MinTTL if you configure CloudFront to forward all headers to your origin (under Headers, if you specify 1 for Quantity and * for Name).

pub fn set_min_ttl(self, input: Option<i64>) -> Self

pub fn default_ttl(self, input: i64) -> Self

This field is deprecated. We recommend that you use the DefaultTTL field in a cache policy instead of this field. For more information, see Creating cache policies or Using the managed cache policies in the Amazon CloudFront Developer Guide.

The default amount of time that you want objects to stay in CloudFront caches before CloudFront forwards another request to your origin to determine whether the object has been updated. The value that you specify applies only when your origin does not add HTTP headers such as Cache-Control max-age, Cache-Control s-maxage, and Expires to objects. For more information, see Managing How Long Content Stays in an Edge Cache (Expiration) in the Amazon CloudFront Developer Guide.

pub fn set_default_ttl(self, input: Option<i64>) -> Self

pub fn max_ttl(self, input: i64) -> Self

This field is deprecated. We recommend that you use the MaxTTL field in a cache policy instead of this field. For more information, see Creating cache policies or Using the managed cache policies in the Amazon CloudFront Developer Guide.

The maximum amount of time that you want objects to stay in CloudFront caches before CloudFront forwards another request to your origin to determine whether the object has been updated. The value that you specify applies only when your origin adds HTTP headers such as Cache-Control max-age, Cache-Control s-maxage, and Expires to objects. For more information, see Managing How Long Content Stays in an Edge Cache (Expiration) in the Amazon CloudFront Developer Guide.

pub fn set_max_ttl(self, input: Option<i64>) -> Self

pub fn build(self) -> DefaultCacheBehavior

Consumes the builder and constructs a DefaultCacheBehavior