Module aws_sdk_cloudfront::operation

Structs

AssociateAlias

Associates an alias (also known as a CNAME or an alternate domain name) with a CloudFront

CreateCachePolicy

Creates a cache policy.

CreateCloudFrontOriginAccessIdentity

Creates a new origin access identity. If you're using Amazon S3 for your origin, you can

CreateDistribution

Creates a new web distribution. You create a CloudFront distribution to tell CloudFront where you

CreateDistributionWithTags

Create a new distribution with tags.

CreateFieldLevelEncryptionConfig

Create a new field-level encryption configuration.

CreateFieldLevelEncryptionProfile

Create a field-level encryption profile.

CreateFunction

Creates a CloudFront function.

CreateInvalidation

Create a new invalidation.

CreateKeyGroup

Creates a key group that you can use with CloudFront signed URLs and signed cookies.

CreateMonitoringSubscription

Enables additional CloudWatch metrics for the specified CloudFront distribution. The

CreateOriginRequestPolicy

Creates an origin request policy.

CreatePublicKey

Uploads a public key to CloudFront that you can use with signed URLs and signed cookies, or with field-level encryption.

CreateRealtimeLogConfig

Creates a real-time log configuration.

CreateStreamingDistribution

This API is deprecated.

CreateStreamingDistributionWithTags

This API is deprecated.

DeleteCachePolicy

Deletes a cache policy.

DeleteCloudFrontOriginAccessIdentity

Delete an origin access identity.

DeleteDistribution

Delete a distribution.

DeleteFieldLevelEncryptionConfig

Remove a field-level encryption configuration.

DeleteFieldLevelEncryptionProfile

Remove a field-level encryption profile.

DeleteFunction

Deletes a CloudFront function.

DeleteKeyGroup

Deletes a key group.

DeleteMonitoringSubscription

Disables additional CloudWatch metrics for the specified CloudFront distribution.

DeleteOriginRequestPolicy

Deletes an origin request policy.

DeletePublicKey

Remove a public key you previously added to CloudFront.

DeleteRealtimeLogConfig

Deletes a real-time log configuration.

DeleteStreamingDistribution

Delete a streaming distribution. To delete an RTMP distribution using the CloudFront API,

DescribeFunction

Gets configuration information and metadata about a CloudFront function, but not the function’s

GetCachePolicy

Gets a cache policy, including the following metadata:

GetCachePolicyConfig

Gets a cache policy configuration.

GetCloudFrontOriginAccessIdentity

Get the information about an origin access identity.

GetCloudFrontOriginAccessIdentityConfig

Get the configuration information about an origin access identity.

GetDistribution

Get the information about a distribution.

GetDistributionConfig

Get the configuration information about a distribution.

GetFieldLevelEncryption

Get the field-level encryption configuration information.

GetFieldLevelEncryptionConfig

Get the field-level encryption configuration information.

GetFieldLevelEncryptionProfile

Get the field-level encryption profile information.

GetFieldLevelEncryptionProfileConfig

Get the field-level encryption profile configuration information.

GetFunction

Gets the code of a CloudFront function. To get configuration information and metadata about

GetInvalidation

Get the information about an invalidation.

GetKeyGroup

Gets a key group, including the date and time when the key group was last modified.

GetKeyGroupConfig

Gets a key group configuration.

GetMonitoringSubscription

Gets information about whether additional CloudWatch metrics are enabled for the specified

GetOriginRequestPolicy

Gets an origin request policy, including the following metadata:

GetOriginRequestPolicyConfig

Gets an origin request policy configuration.

GetPublicKey

Gets a public key.

GetPublicKeyConfig

Gets a public key configuration.

GetRealtimeLogConfig

Gets a real-time log configuration.

GetStreamingDistribution

Gets information about a specified RTMP distribution, including the distribution configuration.

GetStreamingDistributionConfig

Get the configuration information about a streaming distribution.

ListCachePolicies

Gets a list of cache policies.

ListCloudFrontOriginAccessIdentities

Lists origin access identities.

ListConflictingAliases

Gets a list of aliases (also called CNAMEs or alternate domain names) that conflict or

ListDistributions

List CloudFront distributions.

ListDistributionsByCachePolicyId

Gets a list of distribution IDs for distributions that have a cache behavior that’s

ListDistributionsByKeyGroup

Gets a list of distribution IDs for distributions that have a cache behavior that references

ListDistributionsByOriginRequestPolicyId

Gets a list of distribution IDs for distributions that have a cache behavior that’s

ListDistributionsByRealtimeLogConfig

Gets a list of distributions that have a cache behavior that’s associated with the specified

ListDistributionsByWebACLId

List the distributions that are associated with a specified WAF web ACL.

ListFieldLevelEncryptionConfigs

List all field-level encryption configurations that have been created in CloudFront for this account.

ListFieldLevelEncryptionProfiles

Request a list of field-level encryption profiles that have been created in CloudFront for this account.

ListFunctions

Gets a list of all CloudFront functions in your account.

ListInvalidations

Lists invalidation batches.

ListKeyGroups

Gets a list of key groups.

ListOriginRequestPolicies

Gets a list of origin request policies.

ListPublicKeys

List all public keys that have been added to CloudFront for this account.

ListRealtimeLogConfigs

Gets a list of real-time log configurations.

ListStreamingDistributions

List streaming distributions.

ListTagsForResource

List tags for a CloudFront resource.

PublishFunction

Publishes a CloudFront function by copying the function code from the DEVELOPMENT

TagResource

Add tags to a CloudFront resource.

TestFunction

Tests a CloudFront function.

UntagResource

Remove tags from a CloudFront resource.

UpdateCachePolicy

Updates a cache policy configuration.

UpdateCloudFrontOriginAccessIdentity

Update an origin access identity.

UpdateDistribution

Updates the configuration for a web distribution.

UpdateFieldLevelEncryptionConfig

Update a field-level encryption configuration.

UpdateFieldLevelEncryptionProfile

Update a field-level encryption profile.

UpdateFunction

Updates a CloudFront function.

UpdateKeyGroup

Updates a key group.

UpdateOriginRequestPolicy

Updates an origin request policy configuration.

UpdatePublicKey

Update public key information. Note that the only value you can change is the comment.

UpdateRealtimeLogConfig

Updates a real-time log configuration.

UpdateStreamingDistribution

Update a streaming distribution.